+++ WrightTools/artists/__init__.py
@@ -4,4 +4,5 @@
 
 from ._base import *
 from ._colors import *
+from ._helpers import *
 from ._quick import *
