+++ examples/colormaps.py
@@ -4,8 +4,6 @@ Colorbars
 =========
 
 Different colorbars.
-
-Grayscale strategy from http://www.tannerhelland.com/3643/grayscale-image-algorithm-vb6/.
 """
 
 import numpy as np
@@ -16,9 +14,7 @@ from matplotlib import cm
 import WrightTools as wt
 from WrightTools import datasets
 
-
 fig, gs = wt.artists.create_figure(width='double', cols=[1, 1, 'cbar'], nrows=3)
-
 
 p = datasets.COLORS.v2p1_MoS2_TrEE_movie
 data = wt.data.from_COLORS(p, verbose=False)
@@ -28,14 +24,10 @@ data.ai0.symmetric_root(0.5)
 data = data.chop('w1=wm', 'w2', at={'d2': [-600, 'fs']})[0]
 data.ai0.normalize()
 
-
 def fill_row(row, cmap):
     # greyscale
-    rgb = cmap(data.ai0[:])
-    r, g, b, a = [rgb[..., i] for i in range(4)]
     ax = plt.subplot(gs[row, 0])
-    luma = r * 0.3 + g * 0.59 + b * 0.11
-    ax.pcolor(data.w1__e__wm.full, data.w2.full, luma, cmap=cm.gray)
+    ax.pcolor(data, cmap=wt.artists.grayify_cmap(cmap))
     # color
     ax = plt.subplot(gs[row, 1])
     ax.pcolor(data, cmap=cmap)
@@ -43,18 +35,12 @@ def fill_row(row, cmap):
     cax = plt.subplot(gs[row, 2])
     wt.artists.plot_colorbar(cax=cax, label='amplitude', cmap=cmap)
 
-
 cmap = wt.artists.colormaps['default']
 fill_row(0, cmap)
-
-
 cmap = wt.artists.colormaps['wright']
 fill_row(1, cmap)
-
-
 cmap = cm.viridis
 fill_row(2, cmap)
-
 
 # label
 wt.artists.set_fig_labels(xlabel=data.w1__e__wm.label, ylabel=data.w2.label)
